refac: FSD 구조 리팩토링 - entities/features 도메인별 재구조화
- entities를 비즈니스 도메인별로 재구성 (post, guestbook, summary, alarm)
- entities에서 UI 컴포넌트를 features 레이어로 이동
- Notion/OpenAI 인프라 관심사를 도메인별 api 세그먼트로 분리
- Notion 클라이언트 인스턴스를 shared/api로 추출
- theme을 entities에서 features로 이동
- alarm을 features에서 entities로 이동 (독립 도메인)
- 전체 슬라이스 네이밍 단수형 통일 (posts→post, guestbooks→guestbook 등)
- entities→features 상향 import 위반 해소
- vitest에 .claude 디렉토리 exclude 추가
