+++ src/entities/guestbook/model/type.d.ts
