--- src/entities/guestbooks/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./guestbook-card";
