+++ src/entities/post/model/index.ts
