+++ src/entities/post/model/type.d.ts
