+++ src/entities/summary/config/index.ts
