+++ src/features/post/ui/filterable-post.tsx
@@ -2,10 +2,10 @@
 
 import type { DatabaseObjectResponse } from "@notionhq/client/build/src/api-endpoints";
 import { useState } from "react";
-import { Post, Tag } from "@/entities/posts/model";
-import type { TagDatabaseResponse } from "@/entities/posts/model/type";
-import { EmptyPosts } from "@/entities/posts/ui";
-import { TagFilter } from "@/features/tags/ui";
+import { Post, Tag } from "@/entities/post/model";
+import type { TagDatabaseResponse } from "@/entities/post/model/type";
+import { TagFilter } from "@/features/tag/ui";
+import { EmptyPosts } from "./empty-posts";
 import { PostsGrid } from "./posts-grid";
 
 type FilterablePostsProps = {
