+++ src/features/post/ui/index.ts
@@ -1,4 +1,8 @@
 export * from "./client-notion-renderer";
 export * from "./empty-posts";
+export * from "./featured-post";
+export * from "./filterable-post";
 export * from "./post-card";
+export * from "./post-navigator";
+export * from "./posts-grid";
 export * from "./small-post-card";
