--- src/features/posts/ui/index.ts
@@ -1,5 +1,0 @@
-export * from "./ai-summary-button";
-export * from "./featured-post";
-export * from "./filterable-post";
-export * from "./post-navigator";
-export * from "./posts-grid";
