+++ src/features/summary/ui/ai-summary-card.tsx
